Extensions: Incentives. Added ability to use {{customer_id}} code inside incentive descriptions. API changes
