+++ abantecart/abc/extensions/incentive/controllers/storefront/api/account/my_incentives.php
@@ -105,29 +105,6 @@ class ControllerApiAccountMyIncentives extends ASecureControllerAPI
         $this->rest->setResponseData($this->data['incentives']);
         $this->rest->sendResponse(200);
     }
-
-    /**
-     * @param array $params
-     * @return array
-     * @throws InvalidArgumentException
-     * @throws ReflectionException
-     * @throws AException
-     */
-    protected function getMyIncentives($params): array
-    {
-        $params['language_id'] = $params['language_id'] ?: $this->language->getLanguageID();
-        Incentive::setCurrentLanguageID($params['language_id']);
-        $incentives = Incentive::getCustomerIncentives($this->checkout, $params);
-
-        foreach ($incentives as &$incentive) {
-            $incentive = $this->mapIncentiveDataToApiResponse($incentive);
-        }
-
-        return [
-            'total'      => count($incentives),
-            'incentives' => $incentives
-        ];
-    }
 }
